# gadget-store/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Commerce API",
        "description": "API documentation for Laravel Commerce application with product management, sales tracking, categories, and admin authentication",
        "contact": {
            "email": "admin@laravelcommerce.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Retrieve a list of all products (public access)",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Create a new product (requires admin authentication)",
                "operationId": "createProduct",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        },
                                        "created_by_admin": {
                                            "type": "string",
                                            "example": "Admin Name"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Admin authentication required"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "AdminAuth": []
                    }
                ]
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "description": "Retrieve a specific product by its ID (public access)",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "68b74ba7002cda59000d800c"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "description": "Update an existing product (requires admin authentication)",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "68b74ba7002cda59000d800c"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Product data to update",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product updated successfully"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        },
                                        "updated_by_admin": {
                                            "type": "string",
                                            "example": "Admin Name"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Admin authentication required"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "AdminAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "description": "Delete a product (requires admin authentication)",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "68b74ba7002cda59000d800c"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product deleted successfully"
                                        },
                                        "deleted_product": {
                                            "type": "string",
                                            "example": "iPhone 15 Pro"
                                        },
                                        "deleted_by_admin": {
                                            "type": "string",
                                            "example": "Admin Name"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Admin authentication required"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "AdminAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "68b74ba7002cda59000d800c"
                    },
                    "product_name": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "category_id": {
                        "type": "string",
                        "example": "68b74ba7002cda59000d800d",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.990000000000009094947017729282379150390625
                    },
                    "overview": {
                        "type": "string",
                        "example": "Latest iPhone with advanced features",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Detailed product description",
                        "nullable": true
                    },
                    "about": {
                        "type": "string",
                        "example": "About this product",
                        "nullable": true
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "images_url": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "colors": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "what_is_included": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "specification": {
                        "type": "object",
                        "nullable": true
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ProductRequest": {
                "title": "Product Request",
                "description": "Product creation/update request",
                "required": [
                    "product_name",
                    "price",
                    "in_stock"
                ],
                "properties": {
                    "product_name": {
                        "type": "string",
                        "example": "iPhone 15 Pro"
                    },
                    "category_id": {
                        "type": "string",
                        "example": "68b74ba7002cda59000d800d",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.990000000000009094947017729282379150390625
                    },
                    "overview": {
                        "type": "string",
                        "example": "Latest iPhone with advanced features",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Detailed product description",
                        "nullable": true
                    },
                    "about": {
                        "type": "string",
                        "example": "About this product",
                        "nullable": true
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "images_url": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "colors": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "path": {
                                    "type": "string",
                                    "example": "/images/blue.jpg"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Blue"
                                }
                            },
                            "type": "object"
                        },
                        "nullable": true
                    },
                    "what_is_included": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "specification": {
                        "type": "object",
                        "nullable": true
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "AdminAuth": {
                "type": "apiKey",
                "description": "Admin ID for authentication (required for create, update, delete operations)",
                "name": "Admin-ID",
                "in": "header"
            }
        }
    },
...
}